[FIX] Browse user without prefetch when opening database
When running bin/upgrade_odoo, the database is loaded once to fetch the current
database patch level. In the process, the user is browsed and by extension,
the partner_id that it inherits. If the modules that are installed in the
database have added a new field on the user or partner model that has not yet
been instantiated in the database tables, the loading of the database will
fail with an UndefinedColumn exception at that point.

By disabling prefetching of fields, only the columns will be read of the fields
that are actually being requested (in this case lang and tz). This prevents
the ORM from trying to read all the columns that it thinks are in the table,
including the newly added fields.